% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hypothesis.R
\name{linear_hypothesis}
\alias{linear_hypothesis}
\title{Testiranje linearnih hipoteza}
\usage{
linear_hypothesis(model, C, gamma)
}
\arguments{
\item{model}{linearni model za koji se testira}

\item{C}{matrica C}

\item{gamma}{vektor gama}
}
\value{
Ispisuje vrednost F statistike, odgovarajucih stepeni slobode, kao i
  p vrednost testa i vraca listu sa pomenutim vrednostima.
}
\description{
Funkcija za testiranje hipoteza oblika C*beta = gamma
}
